fix/reflow: impl a hack to deal with different indentation depths
In the `reflow_indentations`, the initial reflow is `8`, followed by
indentation depths of `7` which is correct for
```
< 4><3>
      >1<
    ///  A B C
    /// Second
```

now if the first line is too long and is reflown, the reflow start
is going be from column `8` (base 0), the second line is from
column `7` on (which is correct, `4` indent plus `3` for the triple
slash).
Now this information must find it's way through the `Indentation`
obj to the right spot which turns out to be not as trivial as
anticipated, since the space is neither part of the reflown nor
part of the indentation, so we would need to keep track of the fact,
if it's the first line to be reflown or not.

There is big fat XXX TODO FIXME in the code.