Lints and cargo deps
Sadly, I cannot compile this locally (some missing llvm dep, and it doesn't build with no-defaults or with just the `nlprules` feature). So hopefully your CI will catch it (would it be ok to add a simple github action build too, as it will be a common base to ensure it builds for everyone? I could help with that if ok). Anyway, tried to fix as much as i could.

* make all optional dependencies not be auto-added as features -- keeps things tidier
* a few match statement simplifications
* a few "else after return" inlining
* a few `cfg` attributes - still doesn't build, but gets closer
* a tiny spell fix :)